refactor: moved linter integration tests outside unit
